feature-store/requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
docstring_parser==0.16
google-api-core==2.22.0
google-auth==2.36.0
google-cloud-aiplatform==1.71.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.13.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.67.1
grpcio-status==1.67.1
idna==3.10
numpy==2.1.3
packaging==24.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
shapely==2.0.6
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3